Move loan::protocol_tests to baru/tests
By moving them to the crate's `tests` directory we ensure that we're
testing the public API of the library.

We've temporarily made public the `oracle` sub-module in `loan`, but
since it's not fully developed we don't announce it in the changelog.
